Generate a Pandas data frame with entities with a given attribute Fixes #13
